started creating es tests
