+++ backend/backend/settings.py
@@ -48,8 +48,6 @@ INSTALLED_APPS = [
     'rest_framework_social_oauth2',
     'corsheaders',
     'rest_framework',
-    # 'django_elasticsearch_dsl',
-    # 'django_elasticsearch_dsl_drf',
     'social_django',
